+++ src/main/java/org/ioteatime/meonghanyangserver/user/domain/UserEntity.java
@@ -1,0 +1,25 @@
+package org.ioteatime.meonghanyangserver.user.domain;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@NoArgsConstructor
+public class UserEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String nickname;
+
+    @Column(nullable = false)
+    private String email;
+
+    @Column(nullable = false)
+    private String password;
+
+    @Column private String profileImgUrl;
+}
