--- src/main/java/org/ioteatime/meonghanyangserver/user/repository/UserEntity.java
@@ -1,25 +1,0 @@
-package org.ioteatime.meonghanyangserver.user.repository;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@NoArgsConstructor
-public class UserEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String email;
-
-    @Column private String profileImgUrl;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private String nickname;
-}
